Moved commands out of the bin folder
Did not make sense after the move of the bin script
