+++ src/commands/helpers/default-prompt.js
