+++ test/commands/data/invalid-template/package.json
