+++ test/commands/data/valid-template/package.json
