+++ test/commands/data/valid-template/template/package.json
