+++ test/commands/data/valid-template/template/roc.config.js
