+++ test/commands/helpers/data/package/invalid/package.json
