+++ test/commands/helpers/data/package/valid-config/package.json
