+++ test/commands/helpers/data/package/valid-config/roc.config.js
