+++ test/commands/helpers/data/package/valid/package.json
