+++ test/commands/helpers/default-prompt.js
@@ -1,10 +1,10 @@
 import expect from 'expect';
 
 describe('roc', () => {
-    describe('bin', () => {
+    describe('commands', () => {
         describe('helpers', () => {
             describe('defaultPrompt', () => {
-                const prompt = require('../../../src/bin/commands/helpers/default-prompt');
+                const prompt = require('../../../src/commands/helpers/default-prompt');
 
                 it('should have a property named defaultPrompt that is an array', () => {
                     expect(prompt.defaultPrompt).toBeAn(Array);
